Move the command registration into it's own handler
